Competition code from Suffield.
Image recognition: Not implemented, but major progress was made on Saturday. We can now see particles, but need to test movement based on particle recognition. Next update will likely include some form of image processing.

Drive Code:
Driving works, but it needs to be put in robotDrive class instead of in teleopPeriodic. We could also experiment with using FIRST's mecanum drive method...or not. Tower commands all work individually, but the controls either need to be made more intuitive or the drivers need more practice. The pneumatic system needs to not be hotwired, which means re-writing code to use spikes. To use solenoids, the definition is solName = new Solenoid(channel number). Note there is no slot number.